feature/GETP-228: 문자열을 받아 Date 객체 반환하는 유틸리티 함수 구현 및 단위테스트
